third_party/WebKit/LayoutTests/platform/win/external/wpt/referrer-policy/generic/subresource-test/iframe-messaging-expected.txt

This is a testharness.js-based test.
PASS Iframe is responding with HTTP headers 
Harness: the test ran to completion.
